requirements.txt
torch==2.0.0
torchvision==0.15.1
gym==0.26.2
numpy==1.23.5
dm-control==1.0.10
dm-env==1.6
dm-tree==0.1.8
matplotlib==3.7.1
mujoco==2.3.2
scikit-learn==1.2.2
hydra-core==1.3.1
scipy
plotly
tqdm
opencv-python
wandb
imageio
bsuite
ray
